requirements.txt
Babel>=1.3
pbr>=0.5.21,<1.0
# This file is managed by openstack-depends
argparse
cliff>=1.4.3
eventlet>=0.13.0
extras
pecan>=0.2.0
iso8601>=0.1.8
netaddr>=0.7.6
oslo.config>=1.2.0
Paste
PasteDeploy>=1.5.0
Routes>=1.12.3
stevedore>=0.10
WebOb>=1.2.3,<1.3
WSME>=0.5b6
# Optional Stuff that is used by default
alembic>=0.4.1
SQLAlchemy>=0.7.8,<=0.7.99
kombu>=2.4.8

# Identity
python-memcached>=1.48
passlib

pycountry
taskflow
